Move `timezone()` to main module to work around mypy sometimes mixing up the function and module
